Removes listuserscripts alias to fix unloading
bash doesn't transfer aliases into new shells, but module expects them
to be there. This makes unloading a module containing an alias return an
error, which makes a script with -o errexit fail if it tries to unload
said module.

So, I'm just removing the alias from this module (and any future modules
that give me this problem). This one's a particular problem because it's
a default 'core' module, so it's often unloaded.