Standardises all the group-checking code I could find
Where there was a load-check just for a group check, I replaced it with the new "mustBeMemberToLoad" call, and where the load-check contained other code, I used the also-new mustBeMember call.

I have tested many but not all of these, especially the awkward Gaussian one where it uses two groups.