Make vasp module check for alternative archs
As part of the Michael upgrade we now have mix of microarchitectures.  This modified module file checks to see if it is running on Michael and then if it is running on one of the new Cascade Lake nodes and if so selects an AVX512 build of VASP instead of the AVX2 one.

On other clusters it should have no effect.