Closes #87 - Makes Gaussian modules detect small TMPDIRs and create SCRDIR automatically
These changes slightly restructure the logic because it was getting a bit tangled.
With the aid of some new module functions, the checking whether the module is being loaded in a job and checking whether the module has temporary space available are handled separately.

In terms of changes to functionality:

 - previously in a job these modules would always put the GAUSS_SCRDIR in TMPDIR
 - if TMPDIR is small (<1GB) that could lead to problems
 - now they check whether TMPDIR is small and if so, put it in a directory in $HOME/Scratch instead

 - before, the user had to create the GAUSS_SCRDIR
 - this seemed somewhat unnecessary
 - now the module does it instead